Update e3_Basics.inc
More reliable tab select and row select for the Diamond Coins in the list box.   The tabs are not so important as they should be static but looking up tab by name now and saving the index as a variable.  The rows for the diamond coin would have been problematic since the rows are sortable and the row index changes depending on how you have those columns sorted.  We're able to pull the row index from a method on the window object and use that to select the row.