Update e3_Background.inc
Same change as in the main loop but for MQ Next.

Most auto looting is triggered by the EVENT_YourKill routine.  Me.CombatState doesn't always return the correct value so I put an additional check in to look at how many XTarget haters you have.  When that is zero we'll consider that out of combat.

My Wizards seem to have the biggest issue with determining combat state and I've run this through a couple days and it seems much more reliable.

/if (${Auto_Loot} && (${Me.CombatState.NotEqual[Combat]} && ${SpawnCount[xtarhater radius 100]} == 0 || ${combatLooting})) /call loot_Area