Failing behind now reaquires.
With the changes to e3_basic on follow works and leveraging Nav. Now the bot will try the /call AcquireFollow. This will cause you bot to leverage mq2Nav until they get within the required distance. Lets face there are times when bots fall behind and the message is missed. This automates the reacquiring.